fix: APM upgarde
